backend/requirements.txt
fastapi
uvicorn[standard]
beanie
motor
pydantic[email]
python-jose[cryptography]
passlib[bcrypt]
python-dotenv